Surround all dotfile tests with a condition
Add run_tests_on_change.sh to conditionally enable dotfile runtime tests when files change

+++ bin/functions/fsattr.bash
@@ -50,27 +50,30 @@ fsattr() {
       ;;
   esac
 }
-source_relative_once assert.bash
-# fsattr on the fly test suite
-# this setup function takes a PID as an argument and uses it to create a file unique to the process
-# ... Because prior to this I actually had fails here firing up multiple shells at the same time, facepalm.gif
-# Also, if tmp is not part of an "attributable" filesystem (for example, tmpfs is not attributable), this will fail,
-# so all of this will skip if the file system is not attributable.
-# Didn't feel motivated to figure out how to make this work safely outside tmp yet.
-if touch /tmp/attrtest$$ && fsattr /tmp/attrtest$$ test 1 1>/dev/null 2>/dev/null; then
-  fsattr /tmp/attrtest$$ test "" 1>/dev/null 2>/dev/null
-  _fsattr_test_setup() { # $1 is a PID or uniqueness identifier
-    touch /tmp/attr_test$1
-    fsattr /tmp/attr_test$1 a 1
-    fsattr /tmp/attr_test$1 b 2
-  }
-  assert "$(_fsattr_test_setup $$ && fsattr /tmp/attr_test$$ a)" == "1" "setting extended attributes on files should work"
-  assert "$(_fsattr_test_setup $$ && fsattr /tmp/attr_test$$ a 1 && fsattr /tmp/attr_test$$ b 2 && fsattr /tmp/attr_test$$)" == "a=\"1\"\nb=\"2\"" "Listing extended attributes on files should work"
-  assert "$(fsattr /tmp/attr_test$$ a "" && fsattr /tmp/attr_test$$)" == "b=\"2\"" "Deleting a named extended attribute by setting it to blank should work"
-  assert "$(fsattr /tmp/attr_test$$ a 1 && fsattr /tmp/attr_test$$ a "" && fsattr /tmp/attr_test$$ a 1>/dev/null 2>/dev/null; echo $?)" != "0" "Accessing a deleted named extended attribute should fail"
-  [ -f /tmp/attr_test$$ ] && rm /tmp/attr_test$$
-  unset -f _fsattr_test_setup
 
-else
-  echo "Skipping fsattr test suite because /tmp/attrtest$$ is not attributable" 1>&2
+if [ "$RUN_DOTFILE_TESTS" == "true" ]; then
+  source_relative_once assert.bash
+  # fsattr on the fly test suite
+  # this setup function takes a PID as an argument and uses it to create a file unique to the process
+  # ... Because prior to this I actually had fails here firing up multiple shells at the same time, facepalm.gif
+  # Also, if tmp is not part of an "attributable" filesystem (for example, tmpfs is not attributable), this will fail,
+  # so all of this will skip if the file system is not attributable.
+  # Didn't feel motivated to figure out how to make this work safely outside tmp yet.
+  if touch /tmp/attrtest$$ && fsattr /tmp/attrtest$$ test 1 1>/dev/null 2>/dev/null; then
+    fsattr /tmp/attrtest$$ test "" 1>/dev/null 2>/dev/null
+    _fsattr_test_setup() { # $1 is a PID or uniqueness identifier
+      touch /tmp/attr_test$1
+      fsattr /tmp/attr_test$1 a 1
+      fsattr /tmp/attr_test$1 b 2
+    }
+    assert "$(_fsattr_test_setup $$ && fsattr /tmp/attr_test$$ a)" == "1" "setting extended attributes on files should work"
+    assert "$(_fsattr_test_setup $$ && fsattr /tmp/attr_test$$ a 1 && fsattr /tmp/attr_test$$ b 2 && fsattr /tmp/attr_test$$)" == "a=\"1\"\nb=\"2\"" "Listing extended attributes on files should work"
+    assert "$(fsattr /tmp/attr_test$$ a "" && fsattr /tmp/attr_test$$)" == "b=\"2\"" "Deleting a named extended attribute by setting it to blank should work"
+    assert "$(fsattr /tmp/attr_test$$ a 1 && fsattr /tmp/attr_test$$ a "" && fsattr /tmp/attr_test$$ a 1>/dev/null 2>/dev/null; echo $?)" != "0" "Accessing a deleted named extended attribute should fail"
+    [ -f /tmp/attr_test$$ ] && rm /tmp/attr_test$$
+    unset -f _fsattr_test_setup
+
+  else
+    echo "Skipping fsattr test suite because /tmp/attrtest$$ is not attributable" 1>&2
+  fi
 fi
